Extracted GameBoard from BoardPage and reused it for the BoardEditorPage preview
